Fixed length offset (#1)
Default FixedLengthInputFormat doesn't allow for the data to start
after an offset, which is a classical use-case (sound and video files,
scientific formats ...). Also the original reader fails if the last
record read in the file is not exactly of the expected recordsize.

This patch contains a modified copy of the original code allowing
to start reading records after an offset, and to have different
strategies for a last partial record (fail by default as original,
or skip, or fill with a provided byte value).